Guardar ubicacion (falla porque no reocje bien)

+++ app/src/main/java/com/example/practicamoviles_1tr/models/Location.java
@@ -41,7 +41,7 @@ public class Location implements Parcelable {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -49,7 +49,7 @@ public class Location implements Parcelable {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
